more refactoring, deleting old code
