+++ iris-mpc-cpu/src/execution/mod.rs
@@ -1,2 +1,3 @@
+pub mod local;
 pub mod player;
 pub mod session;
