+++ iris-mpc-cpu/src/protocol/mod.rs
@@ -1,3 +1,3 @@
 pub(crate) mod binary;
-pub mod ng_worker;
+pub mod ops;
 pub(crate) mod prf;
